test(jazz): organize query compiler coverage
